refactor(editorIdentity): Move EditorIdentity to its own file
